+++ lib/services/http/tour/tour_utils.dart
